Move header files back to plugins
