+++ Geometry/MTDNumberingBuilder/plugins/CmsMTDConstruction.h
@@ -2,7 +2,7 @@
 #define Geometry_MTDNumberingBuilder_CmsMTDConstruction_H
 #include<string>
 #include<vector>
-#include "Geometry/MTDNumberingBuilder/interface/CmsMTDLevelBuilder.h"
+#include "Geometry/MTDNumberingBuilder/plugins/CmsMTDLevelBuilder.h"
 /**
  * Adds GeometricTimingDets representing final modules to the previous level
  */
