+++ Geometry/TrackerNumberingBuilder/plugins/CmsTrackerDiskBuilder.h
@@ -1,7 +1,7 @@
 #ifndef Geometry_TrackerNumberingBuilder_CmsTrackerDiskBuilder_H
 # define Geometry_TrackerNumberingBuilder_CmsTrackerDiskBuilder_H
 
-# include "Geometry/TrackerNumberingBuilder/interface/CmsTrackerLevelBuilder.h"
+# include "Geometry/TrackerNumberingBuilder/plugins/CmsTrackerLevelBuilder.h"
 # include "FWCore/ParameterSet/interface/types.h"
 # include <string>
